move display formatting logic to proper file
