+++ src/components/dashboard/burn_tokens.tsx
@@ -5,11 +5,9 @@ import {
 	Box,
 	Button,
 	Icon,
-	Input,
 	Link,
 	NumberInput,
 	NumberInputField,
-	Spacer,
 	Stack,
 	Text,
 } from '@chakra-ui/react'
@@ -29,14 +27,10 @@ import {
 	explorerUrl,
 	maxSlippagePercent,
 	maxReserveChangePercent,
-	format,
-	parse,
 	commandTypes,
-	sanitizeNumberInput,
 	curveUtilization,
 	defaultDecimals,
 } from '../../config/constants'
-import { composeQuery } from '../../util/ethers_utils'
 import { CgArrowDownR } from 'react-icons/cg'
 
 import { BigNumber as bignumber } from 'bignumber.js'
@@ -45,7 +39,7 @@ import { Toast } from '../toast'
 import { BiLinkExternal } from 'react-icons/bi'
 import { error_message } from '../../config/error'
 import { isAbleToSendTransaction } from '../../config/validation'
-import { formatBalanceNumber, formatReceiveNumber } from '../../util/display_formatting'
+import { formatBalanceNumber, formatReceiveNumber, format, parse, sanitizeNumberInput } from '../../util/display_formatting'
 
 type mintProps = {
 	dashboardDataSet: any
