+++ src/util/display_formatting.ts
@@ -118,3 +118,13 @@ export function logBigInt(number: String){
 
   return logExponent + logManageableNumber
 }
+
+export const format = (val: any) => val + `%`
+export const parse = (val: any) => val.replace(/^\%/, '')
+export function sanitizeNumberInput(input: any){
+  const sanitizedValue = input
+  .replace(/[-+e]/g, "")
+  .replace(/(\.\d*)\./g, "$1");
+
+  return sanitizedValue
+}
